created new history page and user can navigate from history tab
