+++ src/components/buttons/WindowBack.tsx
@@ -1,13 +1,24 @@
-import { setWindowName } from "../../RTK/slices/toggleWindow";
+import { useNavigate } from "react-router-dom";
 import { useDispatch } from "react-redux";
 
 import NavigateBeforeIcon from "@mui/icons-material/NavigateBefore";
 import IconButton from "@mui/material/IconButton";
+import backToMainTab from "../../utils/tabClose";
 
-function WindowBack() {
+function WindowBack({ page = false }) {
+  const navigate = useNavigate();
   const dispatch = useDispatch();
+
+  const handleBack = () => {
+    if (page) {
+      navigate("/");
+      return;
+    }
+    backToMainTab({ dispatch });
+  };
+
   return (
-    <IconButton title="Back" onClick={() => dispatch(setWindowName(""))}>
+    <IconButton title="Back" onClick={() => handleBack()}>
       <NavigateBeforeIcon />
     </IconButton>
   );
