+++ src/utils/copyToClipboard.ts
@@ -1,7 +1,7 @@
 async function copyToClipboard(text: string) {
   try {
     await navigator.clipboard.writeText(text);
-    console.log("Copying to clipboard was successful!");
+    console.info("Copying to clipboard was successful!");
     return true;
   } catch (err) {
     console.error("Could not copy text: ", err);
